Move tests into correct directory
