+++ services/analyse/tests/Client/EventBridgeEventClientTest.php
@@ -15,23 +15,24 @@ use DateTimeImmutable;
 use Monolog\Test\TestCase;
 use Packages\Event\Enum\Event;
 use Packages\Event\Enum\EventSource;
-use Packages\Event\Model\UploadsFinalised;
+use Packages\Event\Model\CoverageFinalised;
 use Packages\Models\Enum\Environment;
 use Packages\Models\Enum\Provider;
 use PHPUnit\Framework\Attributes\DataProvider;
 
-class EventBridgeEventServiceTest extends TestCase
+class EventBridgeEventClientTest extends TestCase
 {
     #[DataProvider('failedEntryCountDataProvider')]
     public function testPublishEvent(int $failedEntryCount, bool $expectSuccess): void
     {
-        $detail = new UploadsFinalised(
+        $detail = new CoverageFinalised(
             Provider::GITHUB,
             'mock-owner',
             'mock-repository',
             'mock-ref',
             'mock-commit',
             null,
+            99,
             new DateTimeImmutable()
         );
 
@@ -47,8 +48,8 @@ class EventBridgeEventServiceTest extends TestCase
                     'Entries' => [
                         new PutEventsRequestEntry([
                             'EventBusName' => 'mock-event-bus',
-                            'Source' => EventSource::ORCHESTRATOR->value,
-                            'DetailType' => Event::UPLOADS_FINALISED->value,
+                            'Source' => EventSource::ANALYSE->value,
+                            'DetailType' => Event::COVERAGE_FINALISED->value,
                             'Detail' => 'mock-serialized-json'
                         ])
                     ],
